keep src consistent when unmapping
This one is a little convoluted: the symptom is a segfault in free()
when releasing src->buffer. The root cause is that we manage to unmap
a chunk of arena by accident. First we map, then we unmap (correctly),
then malloc reuses the space, then we unmap again using the stale pointer.

It seems too much of bother to track if we should unmap or not, so
let's just clear the pointer when we unmap.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>